product_service/requirements.txt
django
djangorestframework
djangorestframework-simplejwt
gunicorn
mysqlclient
python-decouple